refactor: break createMutationOptions and createSubscriptionOptions into smaller functions
