+++ packages/offix-cache/src/index.ts
@@ -4,3 +4,5 @@ export * from './createMutationOptions';
 export * from "./createOptimisticResponse";
 export * from "./createSubscriptionOptions";
 export * from "./mutationUpdateFunctions";
+export * from "./api/CacheUpdates";
+export * from "./api/CacheOperation";
